Js_assignment_6
Javascript Assignment 6
1) Perform the following operations to provide the implementation for a
Rectangle class. The operations are:
1. Add an area() method to the Rectangle class.
2. Create a Square class that satisfies the following conditions:
○ It is a subclass of Rectangle.
○ It contains a constructor and no other methods.
○ It can use the Rectangle class' area method to print the area
of a Square object.
2) Write a javascript function find_largest to return the nth largest number
in an arrayeg- given an array of integers- [3,45,6,7,23,5,7,8]
find_largest(3) will return third largest number from the above array -
which is 8.
3) Write a JavaScript program which accept a number as input in the
function parameter and insert dashes (-) between each two even
numbers.
For example if you accept 025468 as the output should be 0-254-6-8.
computeDash(025468) -> 0-254-6-8.